requirements.txt
fasttext
pyyaml
pandas